Permission system

+++ admin/models/articles.php
@@ -2,8 +2,8 @@
 if($_GET['action'] == 'add'){
   $title = 'Ajouter un article';
   if(isset($_POST['submit'])){
-    $sql = $bdd->prepare("INSERT INTO articles (title, content, image, publish) VALUES(?, ?, ?, NOW())");
-    if($sql->execute(array($_POST['title'], $_POST['content'], $_POST['image']))){
+    $sql = $bdd->prepare("INSERT INTO articles (title, content, image, publish, owner) VALUES(?, ?, ?, NOW(), ?)");
+    if($sql->execute(array($_POST['title'], $_POST['content'], $_POST['image'], $account['id']))){
       $error = '<div class="alert alert-success">L\'article a bien été ajouté !</div>';
       $_GET['action'] = 'view';
     }else{
@@ -14,7 +14,7 @@ if($_GET['action'] == 'add'){
 if($_GET['action'] == 'edit' && isset($_GET['id'])){
   $title = 'Modifier un article';
   if(isset($_POST['submit'])){
-    $sql = $bdd->prepare("UPDATE articles SET title = ?, content = ?, image = ? WHERE id = ?");
+    $sql = $bdd->prepare("UPDATE articles SET title = ?, content = ?, image = ? WHERE id = ?".($account['admin'] ? "" : " AND owner = ".$account['id']));
     if($sql->execute(array($_POST['title'], $_POST['content'], $_POST['image'], $_GET['id']))){
       $error = '<div class="alert alert-success">L\'article a bien été modifié !</div>';
       $_GET['action'] = 'view';
@@ -25,7 +25,7 @@ if($_GET['action'] == 'edit' && isset($_GET['id'])){
 }
 if($_GET['action'] == 'delete' && isset($_GET['id'])){
   $title = 'Supprimer un article';
-  $sql = $bdd->prepare("DELETE FROM articles WHERE id = ?");
+  $sql = $bdd->prepare("DELETE FROM articles WHERE id = ?".($account['admin'] ? "" : " AND owner = ".$account['id']));
   if($sql->execute(array($_GET['id']))){
     $error = '<div class="alert alert-success">L\'article a bien été supprimé !</div>';
     $_GET['action'] = 'view';
