+++ admin/models/slides.php
@@ -2,8 +2,8 @@
 if($_GET['action'] == 'add'){
   $title = 'Ajouter une slide';
   if(isset($_POST['submit'])){
-    $sql = $bdd->prepare("INSERT INTO slides (name, text, image) VALUES(?, ?, ?)");
-    if($sql->execute(array($_POST['name'], $_POST['text'], $_POST['image']))){
+    $sql = $bdd->prepare("INSERT INTO slides (name, text, image, owner) VALUES(?, ?, ?, ?)");
+    if($sql->execute(array($_POST['name'], $_POST['text'], $_POST['image'], $account['id']))){
       $error = '<div class="alert alert-success">La slide a bien été ajoutée !</div>';
       $_GET['action'] = 'view';
     }else{
@@ -25,7 +25,7 @@ if($_GET['action'] == 'edit' && isset($_GET['id'])){
 }
 if($_GET['action'] == 'delete' && isset($_GET['id'])){
   $title = 'Supprimer une slide';
-  $sql = $bdd->prepare("DELETE FROM slides WHERE id = ?");
+  $sql = $bdd->prepare("DELETE FROM slides WHERE id = ?".($account['admin'] ? "" : " AND owner = ".$account['id']));
   if($sql->execute(array($_GET['id']))){
     $error = '<div class="alert alert-success">La slide a bien été supprimée !</div>';
     $_GET['action'] = 'view';
